chore(antalmanac): housekeeping for data scripts
- Remove stale apps/backend note from fetch-term-data
- Use fs/promises in update-terms
- Replace copy-terms-to-dist with move-terms-to-dist (rename + overwrite)
- Move archiver to devDependencies
