+++ apps/antalmanac/package.json
@@ -15,7 +15,7 @@
         "fetch-term-data": "tsx scripts/fetch-term-data.ts",
         "fetch-departments": "tsx scripts/fetch-departments.ts",
         "get-data": "tsx scripts/fetch-term-data.ts && tsx scripts/fetch-departments.ts && tsx scripts/get-search-data.ts",
-        "copy-terms-to-dist": "tsx scripts/copy-terms-to-dist.ts",
+        "move-terms-to-dist": "tsx scripts/move-terms-to-dist.ts",
         "zip-dist": "tsx scripts/zip-dist.ts",
         "update-terms": "tsx scripts/update-terms.ts",
         "build": "next build",
@@ -44,7 +44,6 @@
         "@trpc/client": "^11.0.0",
         "@trpc/react-query": "^11.0.0",
         "@trpc/server": "^11.0.0",
-        "archiver": "^6.0.2",
         "arctic": "^3.7.0",
         "date-fns": "^2.29.3",
         "dotenv": "^16.4.5",
@@ -93,6 +92,7 @@
         "@types/react-lazyload": "^3.2.0",
         "@types/ua-parser-js": "^0.7.39",
         "@vitejs/plugin-react": "^4.0.4",
+        "archiver": "^6.0.2",
         "aws-sdk": "^2.1550.0",
         "typescript": "^5.7.2",
         "vite": "^4.4.9",
